# add debug log level to help troubleshoot performance issue.