I often see Pull Request tree view and Checkout process on Windows are slower than macOS but have no clue which part is slowing down. Now adding a log level setting to help trouble shoot.

By default the log level is info, the same as before and the output would be
```
[Info] Looking for git repository
[Info] Git repository found, initializing review manager and pr tree view.
[Info] Found GitHub remote
```

Once it's being changed to debug

```
[Debug 438709s 711ms] Looking for git repository
[Debug 438709s 712ms] Git repository found, initializing review manager and pr tree view.
[Debug 438709s 731ms] Found GitHub remote
[Debug 438711s 367ms] Review> no matching pull request metadata found for current branch NewBranch
```

Please note we don't send this info to telemetry as it's for local troubleshooting and users can provide meaningful info if they opt in. Besides, we don't care about date time, only seconds and ms.